better architecture
