+++ crates/lingua/src/processing/mod.rs
@@ -1,6 +1,7 @@
 pub mod adapters;
 pub mod dedup;
 pub mod import;
+mod import_openai_responses;
 pub mod transform;
 
 pub use adapters::{
